Keep Darwin PhotosUI.swift standalone for the IceCubes host script
Move PHPickerCapabilities and the NSDirectionalRectEdge lookalike into
the primary source so xcrun swiftc of PhotosUI.swift still typechecks
without the Linux guest extras.
