+++ full/photosui/PhotosUILookalikes.swift
@@ -77,19 +77,6 @@ open class UIGestureRecognizer: NSObject, @unchecked Sendable {}
 
 open class UITouch: NSObject, @unchecked Sendable {}
 
-public struct NSDirectionalRectEdge: OptionSet, Hashable, Sendable {
-    public let rawValue: UInt
-
-    public init(rawValue: UInt) {
-        self.rawValue = rawValue
-    }
-
-    public static let top = NSDirectionalRectEdge(rawValue: 1 << 0)
-    public static let leading = NSDirectionalRectEdge(rawValue: 1 << 1)
-    public static let bottom = NSDirectionalRectEdge(rawValue: 1 << 2)
-    public static let trailing = NSDirectionalRectEdge(rawValue: 1 << 3)
-    public static let all: NSDirectionalRectEdge = [.top, .leading, .bottom, .trailing]
-}
 #endif
 
 #if os(Linux)
